requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
bleach==6.1.0
blinker==1.8.2
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.1
passlib==1.7.4
pillow==10.3.0
protobuf==5.28.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==25.0.0
python-dotenv==1.0.1
requests==2.32.0
six==1.16.0
tinycss2==1.4.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
webauthn==1.11.1
webencodings==0.5.1
Werkzeug==3.0.6
WTForms==3.2.1
